# api-doc-dist/swagger.json
{
  "components": {
    "schemas": {
...
      "BookingModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tripId": {
            "type": "string"
          },
          "onwerId": {
            "type": "string"
          },
          "memberId": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "attendees": {
            "type": "string"
          },
          "room": {
            "type": "object"
          },
          "addOns": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "guests": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "message": {
            "type": "string"
          },
          "paymentStatus": {
            "type": "string"
          },
          "deposit": {
            "type": "object"
          },
          "discount": {
            "type": "object"
          },
          "status": {
            "type": "string"
          },
          "nextPaymentDate": {
            "type": "string"
          },
          "totalBaseFare": {
            "type": "number"
          },
          "totalAddonFare": {
            "type": "number"
          },
          "discountBaseFare": {
            "type": "number"
          },
          "discountAddonFare": {
            "type": "number"
          },
          "paidAmout": {
            "type": "number"
          },
          "totalFare": {
            "type": "number"
          },
          "pendingAmout": {
            "type": "number"
          },
          "paymentHistory": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "createdAt": {
            "type": "number",
            "format": "unix timestamp"
          },
          "updatedAt": {
            "type": "number",
            "format": "unix timestamp"
          }
        }
      },
...
      "HostRequestModel": {
        "type": "object",
        "properties": {
          "awsUserId": {
            "type": "string"
          },
          "kindOfTripHostingOverview": {
            "type": "string",
            "description": "Describe type of trip you planning to host"
          },
          "hostingCapacity": {
            "type": "string",
            "enum": ["1", "2-3", "5-10", "10+"]
          },
          "hostingForCountries": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "targettingTypesOfTravelers": {
            "type": "string",
            "description": "Hosting trip for kind of audience group?",
            "enum": [true, false]
          },
          "isActive": {
            "type": "boolean",
            "description": "If `isActive=true` host request is not deleted",
            "enum": [true, false]
          },
          "groupTripHostingExperience": {
            "type": "string",
            "description": "Host trip hosting experience",
            "enum": ["professional", "intermediate", "beginner"]
          },
          "pastAccomplishmentReferences": {
            "type": "string",
            "description": "Host past experiences reference links"
          },
          "status": {
            "type": "boolean",
            "description": "Request status",
            "enum": ["pending", "approved", "declined"]
          },
          "updatedAt": {
            "type": "number",
            "format": "Unix timestamp"
          },
          "createdAt": {
            "type": "number",
            "format": "Unix timestamp"
          }
        }
      },
...
      "CommonErrorFormat": {
        "type": "object",
        "required": ["status", "result"],
        "properties": {
          "status": {
            "type": "string",
            "default": "error"
          },
          "result": {
            "type": "object",
            "required": ["errorCode", "message"],
            "properties": {
              "errorCode": {
                "type": "string",
                "description": "Error code to handle internationalization(i18n)"
              },
              "message": {
                "type": "string",
                "description": "Error message in english"
              },
              "field": {
                "type": "string",
                "description": "Error field name"
              }
            }
          }
        }
      },
      "CommonSuccessFormat": {
        "type": "object",
        "required": ["status", "result"],
        "properties": {
          "status": {
            "type": "string",
            "default": "success"
          },
          "result": {
            "type": "string",
            "default": "success"
          }
        }
      },
...
      "bookingSchema": {
        "properties": {
          "tripId": {
            "type": "string"
          },
          "stripePaymentMethod": {
            "type": "object"
          },
          "currency": {
            "type": "string"
          },
          "attendees": {
            "type": "number"
          },
          "room": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "cost": {
                "type": "number"
              },
              "available": {
                "type": "number"
              }
            }
          },
          "addOns": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "cost": {
                  "type": "number"
                },
                "available": {
                  "type": "number"
                }
              }
            }
          },
          "paymentStatus": {
            "type": "string",
            "enum": ["full", "deposit", "payasyougo"]
          },
          "message": {
            "type": "string"
          },
          "deposit": {
            "type": "object",
            "properties": {
              "amount": {
                "type": "number"
              },
              "expirationDate": {
                "type": "number"
              },
              "includeAddOns": {
                "type": "boolean"
              }
            }
          },
          "discount": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "discType": {
                "type": "string",
                "enum": ["amount", "percentage"]
              },
              "amount": {
                "type": "number"
              },
              "expirationDate": {
                "type": "number"
              },
              "includeAddOns": {
                "type": "number"
              }
            }
          },
          "guests": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "relationship": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    },
    "parameters": {
...
      "pathHostRequestId": {
        "name": "hostRequestId",
        "in": "path",
        "description": "Id of the host request",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "pathBookingId": {
        "name": "bookingId",
        "in": "path",
        "description": "Id of the booking request",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "queryPage": {
        "name": "page",
        "in": "query",
        "description": "Page number of pagination",
        "schema": {
          "type": "number",
          "default": 0
        }
      },
...
      "queryLimit": {
        "name": "limit",
        "in": "query",
        "description": "Limit of the records on a each page",
        "schema": {
          "type": "number",
          "default": 25,
          "minimum": 25,
          "maximum": 100
        }
      },
      "querySortBy": {
        "name": "sortBy",
        "in": "query",
        "description": "Sort by",
        "schema": {
          "type": "string",
          "default": "updatedAt",
          "enum": ["updatedAt"]
        }
      },
...
      "querySortOrder": {
        "name": "sortOrder",
        "in": "query",
        "description": "Sort order applied on `sortBy` fields",
        "schema": {
          "type": "number",
          "enum": [1, -1],
          "default": -1
        }
      },
...
      "queryTripId": {
        "name": "tripId",
        "in": "query",
        "description": "Trip Id",
        "schema": {
          "type": "string"
        }
      },
      "queryIsHost": {
        "name": "isHost",
        "in": "query",
        "description": "This will list the details as per the host permissions else it will list the details as normal user.",
        "schema": {
          "type": "boolean",
          "enum": [true, false]
        }
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    },
    "requestBodies": {
      "createHostRequest": {
        "description": "Create host request payload",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "properties": {
                "kindOfTripHostingOverview": {
                  "type": "string",
                  "description": "Describe type of trip you planning to host"
                },
                "hostingCapacity": {
                  "type": "string",
                  "enum": ["1", "2-3", "5-10", "10+"]
                },
                "hostingForCountries": {
                  "type": "array"
                },
                "targettingTypesOfTravelers": {
                  "type": "string",
                  "description": "Hosting trip for kind of audience group?"
                },
                "groupTripHostingExperience": {
                  "type": "string",
                  "description": "Host trip hosting experience",
                  "enum": ["professional", "intermediate", "beginner"]
                },
                "pastAccomplishmentReferences": {
                  "type": "string",
                  "description": "Host past experiences reference links"
                }
              }
            }
          }
        }
      },
      "updateHostRequest": {
        "description": "Update host request payload",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "properties": {
                "action": {
                  "type": "string",
                  "enum": ["approved", "declined"]
                }
              }
            }
          }
        }
      },
      "createBooking": {
        "description": "Create new booking payload",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/bookingSchema"
            }
          }
        }
      },
      "updateBooking": {
        "description": "Create new booking payload",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/bookingSchema"
            }
          }
        }
      },
...
    }
  },
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0",
    "title": "Tripsha API documentation",
    "description": "This is complete API documentation of all the API",
    "termsOfService": "https://tripsha.com/terms/",
    "contact": {
      "name": "Tripsha API Team",
      "email": "madanesanjay@yahoo.com",
      "url": "https://tripsha.com"
    }
  },
  "servers": [
    {
      "url": "https://staging-api.tripsha.com"
    }
  ],
  "paths": {
...
    "/host-requests": {
      "get": {
        "description": "Return list of matching host requests",
        "operationId": "listHostRequests",
        "tags": ["HostRequests"],
        "parameters": [
          {
            "$ref": "#/components/parameters/queryPage"
          },
          {
            "$ref": "#/components/parameters/queryLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "Host request response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "result"],
                  "properties": {
                    "status": {
                      "type": "string",
                      "default": "success"
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/HostRequestModel"
                          }
                        },
                        "count": {
                          "type": "number",
                          "description": "Count of the current records"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Creates a host request",
        "operationId": "createHostRequest",
        "tags": ["HostRequests"],
        "requestBody": {
          "$ref": "#/components/requestBodies/createHostRequest"
        },
        "responses": {
          "200": {
            "description": "Create host request response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "result"],
                  "properties": {
                    "status": {
                      "type": "string",
                      "default": "success"
                    },
                    "result": {
                      "type": "string",
                      "default": "request id"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorFormat"
                }
              }
            }
          }
        }
      }
    },
    "/host-requests/{hostRequestId}": {
      "get": {
        "description": "Returns a host request based on a request ID",
        "operationId": "getHostRequest",
        "tags": ["HostRequests"],
        "parameters": [
          {
            "$ref": "#/components/parameters/pathHostRequestId"
          }
        ],
        "responses": {
          "200": {
            "description": "Host Request response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "result"],
                  "properties": {
                    "status": {
                      "type": "string",
                      "default": "success"
                    },
                    "result": {
                      "$ref": "#/components/schemas/HostRequestModel"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorFormat"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Updates a single host request based on the ID supplied, also it will update the status if request",
        "operationId": "updateHostRequest",
        "tags": ["HostRequests"],
        "parameters": [
          {
            "$ref": "#/components/parameters/pathHostRequestId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/updateHostRequest"
        },
        "responses": {
          "204": {
            "description": "Host request updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonSuccessFormat"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorFormat"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "Deletes a single host request based on the ID supplied",
        "operationId": "deleteHostRequest",
        "tags": ["HostRequests"],
        "parameters": [
          {
            "$ref": "#/components/parameters/pathHostRequestId"
          }
        ],
        "responses": {
          "204": {
            "description": "Host request deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonSuccessFormat"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorFormat"
                }
              }
            }
          }
        }
      }
    },
    "/bookings": {
      "get": {
        "description": "Returns list of bookings related to logged in user.",
        "operationId": "listBookings",
        "tags": ["Bookings"],
        "parameters": [
          {
            "$ref": "#/components/parameters/queryPage"
          },
          {
            "$ref": "#/components/parameters/queryLimit"
          },
          {
            "$ref": "#/components/parameters/querySortBy"
          },
          {
            "$ref": "#/components/parameters/querySortOrder"
          },
          {
            "$ref": "#/components/parameters/queryTripId"
          },
          {
            "$ref": "#/components/parameters/queryIsHost"
          }
        ],
        "responses": {
          "200": {
            "description": "Booking list response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "result"],
                  "properties": {
                    "status": {
                      "type": "string",
                      "default": "success"
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "count": {
                          "type": "number",
                          "description": "Current records count"
                        },
                        "total": {
                          "type": "number",
                          "description": "Total number of records."
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/BookingModel"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Creates a booking",
        "operationId": "createBooking",
        "tags": ["Bookings"],
        "requestBody": {
          "$ref": "#/components/requestBodies/createBooking"
        },
        "responses": {
          "200": {
            "description": "Booking response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "result": {
                      "type": "string",
                      "description": "Booking id"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorFormat"
                }
              }
            }
          }
        }
      }
    },
    "/bookings/{bookingId}": {
      "get": {
        "description": "Returns a booking based on a booking ID",
        "operationId": "findBookingById",
        "tags": ["Bookings"],
        "parameters": [
          {
            "$ref": "#/components/parameters/pathBookingId"
          }
        ],
        "responses": {
          "200": {
            "description": "Booking response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "result"],
                  "properties": {
                    "status": {
                      "type": "string",
                      "default": "success"
                    },
                    "result": {
                      "$ref": "#/components/schemas/BookingModel"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Updates a single booking based on the ID supplied and data",
        "operationId": "updateBooking",
        "tags": ["Bookings"],
        "parameters": [
          {
            "$ref": "#/components/parameters/pathBookingId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/updateBooking"
        },
        "responses": {
          "204": {
            "description": "Booking updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonSuccessFormat"
                }
              }
            }
          },
          "default": {
            "description": "Error Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorFormat"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "deletes a single booking based on the ID supplied",
        "operationId": "deleteBooking",
        "tags": ["Bookings"],
        "parameters": [
          {
            "$ref": "#/components/parameters/pathBookingId"
          }
        ],
        "responses": {
          "204": {
            "description": "Booking deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonSuccessFormat"
                }
              }
            }
          },
          "default": {
            "description": "Error Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorFormat"
                }
              }
            }
          }
        }
      }
    },
...
  }
}